优化

+++ src/main/java/com/ljd/Food_Delivery/domain/entity/AddressBookEntity.java
@@ -42,7 +42,7 @@ public class AddressBookEntity extends BaseEntity{
 
     private String label;
 
-    private byte isDefault;
+    private Integer isDefault;
 
     private int isDeleted;
 }
